[MODEL-24701] -- refactor
